refactor: fix to run jest
