+++ package.json
@@ -7,7 +7,9 @@
   "scripts": {
     "dev": "plasmo dev",
     "build": "plasmo build",
-    "package": "plasmo package"
+    "package": "plasmo package",
+    "test": "jest",
+    "test:watch": "jest --watch"
   },
   "dependencies": {
     "@plasmohq/messaging": "^0.6.2",
@@ -19,11 +21,16 @@
   },
   "devDependencies": {
     "@ianvs/prettier-plugin-sort-imports": "4.1.1",
+    "@jest/globals": "^29.7.0",
     "@types/chrome": "0.0.258",
+    "@types/jest": "^29.5.12",
     "@types/node": "20.11.5",
     "@types/react": "18.2.48",
     "@types/react-dom": "18.2.18",
+    "jest": "^29.7.0",
     "prettier": "3.2.4",
+    "ts-jest": "^29.1.5",
+    "ts-node": "^10.9.2",
     "typescript": "5.3.3"
   },
   "manifest": {
